@@ -3,21 +3,26 @@ name = "callsight-api"
 version = "0.1.0"
 description = "Add your description here"
 readme = "README.md"
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 dependencies = [
     "assemblyai>=0.40.2",
     "azure-ai-language-conversations>=1.1.0",
     "azure-ai-textanalytics>=5.3.0",
     "azure-core>=1.34.0",
+    "azure-storage-blob>=12.25.1",
+    "boto3>=1.38.17",
+    "chainlit>=2.5.5",
     "email-validator>=2.2.0",
     "fastapi>=0.115.12",
     "librosa>=0.11.0",
+    "msal>=1.32.3",
     "openai>=1.78.1",
     "pre-commit>=4.2.0",
     "pydantic>=2.11.4",
     "pydantic-core>=2.33.2",
     "pydantic-settings>=2.9.1",
-    "python-multipart>=0.0.20",
+    "python-multipart>=0.0.18",
+    "reportlab>=4.4.1",
     "ruff>=0.11.9",
     "supabase>=2.15.1",
     "tiktoken>=0.9.0",
