@@ -6,6 +6,7 @@ from app.services.embeddings import (
     chat_with_context,
     suggestions_with_context,
     get_last_embeddings,
+    chat_with_specific_transcript,
 )
 from app.core.config import settings
 import os
@@ -41,9 +42,12 @@ def needs_context(prompt: str, previous_response_id: str = None):
                 "content": "You're an assistant helping decide whether a user's question "
                 "requires searching previous customer service transcripts. "
                 "If the question is about a person, event, or conversation that may have occurred "
-                "in a past interaction, reply with 'needs context'. If it's a general question "
+                "in a past interaction you do not have access to, reply with 'needs context'. If it's a general question "
                 "that can be answered without having to retrieve any more past data, reply with 'no context'. "
-                "Only reply with one of these two phrases.",
+                "Only reply with one of these two phrases."
+                "Before replying, consider also the context you already have on this conversation from previous responses."
+                "If the user asked you a question before, and then asks a follow question which can be contested with the context you have as of now, reply 'no context'."
+                "If using the previous response id you see that the user previously asked a question, regardless of wether it needed context, reply 'no context' if you can answer this new question using the information you already had in this chatbot conversation.",
             },
             {"role": "user", "content": prompt},
         ],
@@ -85,6 +89,59 @@ def create_messages_for_supabase(
     return user_message, chatbot_message
 
 
+async def send_to_responses_api(
+    gpt_model: str,
+    system_message: str,
+    user_message: str,
+    previous_response_id: str = None,
+):
+    if system_message is None or len(system_message) == 0:
+        input_messages = [{"role": "user", "content": user_message}]
+    else:
+        input_messages = [
+            {"role": "system", "content": system_message},
+            {"role": "user", "content": user_message},
+        ]
+
+    response = client.responses.create(
+        model=gpt_model, previous_response_id=previous_response_id, input=input_messages
+    )
+
+    return response
+
+
+async def save_messages_to_supabase(
+    conversation_id: str,
+    user_message: str,
+    response_id: str,
+    response_output_text: str,
+    response_created_at: str,
+    supabase: Client,
+    prev_response_id: str = None,
+):
+    user_message, chatbot_message = create_messages_for_supabase(
+        conversation_id,
+        user_message,
+        response_id,
+        response_output_text,
+        response_created_at,
+        prev_response_id,
+    )
+    supabase.table("chatbot_messages").insert(user_message).execute()
+    supabase.table("chatbot_messages").insert(chatbot_message).execute()
+
+
+async def generate_title(gpt_model: str, response_id: str):
+    title = client.responses.create(
+        model=gpt_model,
+        previous_response_id=response_id,
+        input=[
+            {"role": "user", "content": "Crea un titulo para esta conversación"},
+        ],
+    )
+    return title
+
+
 @router.post("/chat")
 async def post_chat(
     request: ChatRequest,
@@ -107,23 +164,12 @@ async def post_chat(
         else:
             message = request.prompt
 
-        response = client.responses.create(
-            model=GPT_MODEL,
-            input=[
-                {
-                    "role": "system",
-                    "content": "You answer questions for agents of a call center working with multiple companies.",
-                },
-                {"role": "user", "content": message},
-            ],
-        )
-        title = client.responses.create(
-            model=gpt_model,
-            previous_response_id=response.id,
-            input=[
-                {"role": "user", "content": "Crea un titulo para esta conversación"},
-            ],
-        )
+        system_message = "You answer questions for agents of a call center working with multiple companies. YOU NEVER USE BLOCKS OF CODE OR BLOCK CODES. You may or may not use markdown, but you never use blocks of code regardless"
+        response = await send_to_responses_api(gpt_model, system_message, message)
+
+        title = await generate_title(gpt_model, response.id)
+
+        # used for storing in our supabase table
         chatbot_conversation = {
             "chatbot_conversation_id": conversation_id,
             "user_id": current_user.id,
@@ -131,18 +177,16 @@ async def post_chat(
             "last_response_id": response.id,
             "model": gpt_model,
         }
+        supabase.table("chatbot_conversations").insert(chatbot_conversation).execute()
 
-        user_message, chatbot_message = create_messages_for_supabase(
+        await save_messages_to_supabase(
             conversation_id,
             request.prompt,
             response.id,
             response.output_text,
             response.created_at,
+            supabase,
         )
-
-        supabase.table("chatbot_conversations").insert(chatbot_conversation).execute()
-        supabase.table("chatbot_messages").insert(user_message).execute()
-        supabase.table("chatbot_messages").insert(chatbot_message).execute()
 
         return {
             "response": response.output_text,
@@ -185,34 +229,26 @@ async def continue_chat(
         else:
             message = request.prompt
 
-        response = client.responses.create(
-            model=GPT_MODEL,
-            previous_response_id=previous_response_id,
-            input=[
-                {"role": "system", "content": "Eres un asistente para un call center."},
-                {"role": "user", "content": message},
-            ],
-        )
-
-        # use the obtained info to generate supabase appropiate objects
-        user_message, chatbot_message = create_messages_for_supabase(
-            conversation_id,
-            request.prompt,
-            response.id,
-            response.output_text,
-            response.created_at,
-            previous_response_id,
+        system_message = "Eres un asistente para un call center. YOU NEVER USE BLOCKS OF CODE OR BLOCK CODES. You may or may not use markdown, but you never use blocks of code regardless"
+        response = await send_to_responses_api(
+            GPT_MODEL, system_message, message, previous_response_id
         )
 
         # modify the table to change the last message
         supabase.table("chatbot_conversations").update(
             {"last_response_id": response.id}
-        ).eq(
-            "chatbot_conversation_id", conversation_id
-        ).execute()  # modify the table to change the last message
+        ).eq("chatbot_conversation_id", conversation_id).execute()
 
-        supabase.table("chatbot_messages").insert(user_message).execute()
-        supabase.table("chatbot_messages").insert(chatbot_message).execute()
+        await save_messages_to_supabase(
+            conversation_id,
+            request.prompt,
+            response.id,
+            response.output_text,
+            response.created_at,
+            supabase,
+            previous_response_id,
+        )
+
         return {"response": response.output_text}
     except Exception as e:
         raise HTTPException(status_code=500, detail=str(e))
@@ -269,7 +305,9 @@ async def get_all_chats(
         )
 
         if response.data is None or len(response.data) == 0:
-            raise HTTPException(status_code=404, detail="User profile not found")
+            raise HTTPException(
+                status_code=404, detail="There are no conversations saved"
+            )
 
         return response.data
     except HTTPException:
@@ -318,5 +356,114 @@ async def get_chat_history(
         return response.data
     except HTTPException:
         raise
+    except Exception as e:
+        raise HTTPException(status_code=500, detail=str(e))
+
+
+@router.post("/continue/specific/{previous_response_id}")
+async def chat_with_specific_call(
+    request: ChatRequest,
+    specific_conversation_id: str,
+    previous_response_id: str,
+    current_user=Depends(get_current_user),
+    supabase: Client = Depends(get_supabase),
+    token_budget: int = 4096 - 500,
+):
+    """
+    Continuar un chat existente
+    """
+    try:
+        if needs_context(request.prompt, previous_response_id):
+            message = await chat_with_specific_transcript(
+                conversation_id=specific_conversation_id,
+                current_user=current_user,
+                supabase=supabase,
+                query=request.prompt,
+                token_budget=token_budget,
+            )
+        else:
+            message = request.prompt
+
+        system_message = "Eres un asistente para un call center. YOU NEVER USE BLOCKS OF CODE OR BLOCK CODES. You may or may not use markdown, but you never use blocks of code regardless"
+        response = await send_to_responses_api(
+            GPT_MODEL, system_message, message, previous_response_id
+        )
+
+        # modify the table to change the last message
+        """ supabase.table("chatbot_conversations").update(
+            {"last_response_id": response.id}
+        ).eq("chatbot_conversation_id", conversation_id).execute()
+
+        await save_messages_to_supabase(
+            conversation_id,
+            request.prompt,
+            response.id,
+            response.output_text,
+            response.created_at,
+            supabase,
+            previous_response_id,
+        ) """
+
+        return {
+            "response": response.output_text,
+            "id": response.id,
+        }
+    except Exception as e:
+        raise HTTPException(status_code=500, detail=str(e))
+
+
+@router.post("/chat/specific")
+async def post_chat_specific_call(
+    specific_conversation_id: str,
+    request: ChatRequest,
+    current_user=Depends(get_current_user),
+    supabase: Client = Depends(get_supabase),
+    token_budget: int = 4096 - 500,
+):
+    """Create a new chat"""
+    try:
+        gpt_model = GPT_MODEL
+        # if the question needs previous context, start a new chat with added embeddings
+        if needs_context(request.prompt):
+            message = await chat_with_specific_transcript(
+                conversation_id=specific_conversation_id,
+                current_user=current_user,
+                supabase=supabase,
+                query=request.prompt,
+                token_budget=token_budget,
+            )
+        else:
+            message = request.prompt
+
+        system_message = "You answer questions for agents of a call center working with multiple companies. YOU NEVER USE BLOCKS OF CODE OR BLOCK CODES. You may or may not use markdown, but you never use blocks of code regardless."
+        response = await send_to_responses_api(gpt_model, system_message, message)
+
+        title = await generate_title(gpt_model, response.id)
+
+        # used for storing in our supabase table
+        """ chatbot_conversation = {
+            "chatbot_conversation_id": conversation_id,
+            "user_id": current_user.id,
+            "title": title.output_text,
+            "last_response_id": response.id,
+            "model": gpt_model,
+        }
+        supabase.table("chatbot_conversations").insert(chatbot_conversation).execute()
+
+        await save_messages_to_supabase(
+            conversation_id,
+            request.prompt,
+            response.id,
+            response.output_text,
+            response.created_at,
+            supabase,
+        ) """
+
+        """ "conversation_id": conversation_id, """
+        return {
+            "response": response.output_text,
+            "title": title.output_text,
+            "id": response.id,
+        }
     except Exception as e:
         raise HTTPException(status_code=500, detail=str(e))
