@@ -1,13 +1,19 @@
 from fastapi import APIRouter, Depends, HTTPException
+from pydantic import BaseModel
 from supabase import Client
 
 from app.db.session import get_supabase
-from app.api.routes.auth import check_admin_role
 
 router = APIRouter(prefix="/companies", tags=["companies"])
 
 
-@router.get("/", dependencies=[Depends(check_admin_role)])
+class Company(BaseModel):
+    name: str
+    logo: str
+    category: str
+
+
+@router.get("/")
 async def get_all_companies(supabase: Client = Depends(get_supabase)):
     try:
         response = supabase.table("company_client").select("*").execute()
@@ -16,6 +22,7 @@ async def get_all_companies(supabase: Client = Depends(get_supabase)):
         raise HTTPException(status_code=500, detail=str(e))
 
 
+# Returns the company name and number of users of each company
 @router.get("/companySize")
 async def get_company_size(supabase: Client = Depends(get_supabase)):
     try:
@@ -33,6 +40,7 @@ async def get_company_size(supabase: Client = Depends(get_supabase)):
         raise HTTPException(status_code=500, detail=str(e))
 
 
+# Returns the users of a given company
 @router.get("/{name}/list")
 async def get_users_in_company(name: str, supabase: Client = Depends(get_supabase)):
     try:
@@ -54,4 +62,36 @@ async def get_users_in_company(name: str, supabase: Client = Depends(get_supabas
 
         return {"participants": users_response.data}
     except Exception as e:
+        raise HTTPException(status_code=500, detail=str(e))
+
+
+@router.post("/create")
+async def create_company(company: Company, supabase: Client = Depends(get_supabase)):
+    try:
+        responseCategory = (
+            supabase.table("category")
+            .select("category_id, name")
+            .eq("name", company.category)
+            .execute()
+        )
+        print(responseCategory.data[0])
+        print(responseCategory.data[0]["category_id"])
+
+        if len(responseCategory.data) == 0:
+            raise HTTPException(status_code=404, detail="Category not found")
+
+        print(company.name)
+
+        company_data = {
+            "name": company.name,
+            "logo": company.logo,
+            "category_id": responseCategory.data[0]["category_id"],
+        }
+
+        print(company_data)
+
+        response = supabase.table("company_client").insert(company_data).execute()
+        return {"message": "Company created successfully", "company": response.data}
+    except Exception as e:
+        print(e)
         raise HTTPException(status_code=500, detail=str(e))
